fix(tui): anchor diff file menus to the screen
